idc/requirements/common.txt
Django==1.9.2
psycopg2==2.6.1
django-froala-editor==2.0.1
django-mptt==0.8.0
django-mptt-admin==0.3.2
django-simple-history==1.8.0
Pillow==3.1.0
